+++ packages/helux-core/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @helux/core
 
+## 4.3.2
+
+### Patch Changes
+
+- 94cd307: build(4.3.2): optimize delFnDepData
+  - @helux/hooks-impl@4.3.2
+  - @helux/types@4.3.2
+  - @helux/utils@4.3.2
+
+## 4.3.1
+
+### Patch Changes
+
+- 41b9f1d: build(4.3.1): support mutate cancel
+  - @helux/hooks-impl@4.3.1
+  - @helux/types@4.3.1
+  - @helux/utils@4.3.1
+
+## 4.3.0
+
+### Minor Changes
+
+- ac52a46: build(4.2.8): fix unwatch
+- 5520458: build(4.2.7): fix unwatch
+
+### Patch Changes
+
+- Updated dependencies [ac52a46]
+- Updated dependencies [5520458]
+  - @helux/hooks-impl@4.3.0
+  - @helux/types@4.3.0
+  - @helux/utils@4.3.0
+
 ## 4.2.7
 
 ### Patch Changes