+++ packages/helux-openinula/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @helux/openinula
 
+## 4.3.2
+
+### Patch Changes
+
+- Updated dependencies [94cd307]
+  - @helux/core@4.3.2
+
+## 4.3.1
+
+### Patch Changes
+
+- Updated dependencies [41b9f1d]
+  - @helux/core@4.3.1
+
+## 4.3.0
+
+### Minor Changes
+
+- ac52a46: build(4.2.8): fix unwatch
+- 5520458: build(4.2.7): fix unwatch
+
+### Patch Changes
+
+- Updated dependencies [ac52a46]
+- Updated dependencies [5520458]
+  - @helux/core@4.3.0
+
 ## 4.2.7
 
 ### Patch Changes